[FIX] web: do not hide button in pdf viewer in some cases
The pdf viewer uses an external library to display a pdf file.  This
library displays a few controls, such as a 'Download' button.  However,
this button is sometimes displayed in the top bar, sometimes in the top
right dropdown menu (depending on the size of the widget).

For some unknown reason, the pdf viewer widget disabled the download button.
But this was inconsistent: the download button was only hidden when it
was in the tob bar, and not in the dropdown menu.  With this commit, we
simply remove this behaviour.