[IMP] sale_timesheet: changing ordered qty update planned hours on task
When changing the ordered quantity on a Sales order line, the
planned hours of its related task will be updated too.
This is obviously applicable for SO line that have created
a task at SO confirmation.
If the SO is cancel, or put in draft state, updating the ordered
quantity will change the planned hours of the task. This operation
is however prevent on locked SO, since ordered quantity is a
protected field.

Task #1856984