[FIX] web: increase timeout for warning in logs
Before this commit, it could happen quite frequently that the runbot
displayed a warning (coming from the JS) when some tests/tours were
executed.  This is the standard warning that some modules are not yet
loaded.  This is not harmful in anyway, but is visible in the runbot
logs.

With this commit, we increase the timeout to 6s.  This is not really a
good sign, hopefully cold starts for building assets will not
increase again in the future.