[IMP] purchase: Extract code from _compute_amount to allow inheritance
Purpose
=======

Currently the discount mechanism is not implemented in the purchase module.
This limitation is could be avoided if the _compute_amount arguments could
be overwritten.

While waiting the feature to be implemented in master, we simply introduce
a hook method to allow an easy inheritance of the argument values.

Closes #11235

PS: Also move compute methods under the fields definition to follow
    the coding guidelines